Refactor backend

+++ backend/CaffShop/Services/CaffItemService.cs
@@ -1,8 +1,13 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using CaffShop.Entities;
 using CaffShop.Interfaces;
+using CaffShop.Models.Exceptions;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.StaticFiles;
 using Microsoft.EntityFrameworkCore;
 
 namespace CaffShop.Services
@@ -28,7 +33,14 @@ namespace CaffShop.Services
             if (withOwner)
                 q = q.Include(nameof(CaffItem.Owner));
 
-            return await q.FirstOrDefaultAsync(i => i.Id == id);
+            try
+            {
+                return await q.FirstAsync(i => i.Id == id);
+            }
+            catch (InvalidOperationException)
+            {
+                throw new CaffItemNotFoundException();
+            }
         }
 
         public async Task<IEnumerable<CaffItem>> GetCaffItems(bool withOwner = false)
@@ -54,8 +66,12 @@ namespace CaffShop.Services
             return await q.ToListAsync();
         }
 
-        public async Task DeleteCaffItem(CaffItem item)
+        public async Task DeleteCaffItem(CaffItem item, long userId, bool isAdmin)
         {
+            // Admins can delete all items. Users are allowed to delete only their own items.
+            if (item.OwnerId != userId && !isAdmin)
+                throw new UserNotAllowedToDeleteCaffException();
+
             _context.CaffItems.Remove(item);
             await _context.SaveChangesAsync();
         }
@@ -65,6 +81,39 @@ namespace CaffShop.Services
             await _context.CaffItems.AddAsync(item);
             await _context.SaveChangesAsync();
             return item;
+        }
+
+        public void DeleteCaffItemFromDisc(CaffItem item)
+        {
+            if (System.IO.File.Exists(item.CaffPath))
+                System.IO.File.Delete(item.CaffPath);
+
+            if (System.IO.File.Exists(item.PreviewPath))
+                System.IO.File.Delete(item.PreviewPath);
+        }
+
+        // Returns a filestream
+        public async Task<FileStreamResult> GetPreviewImage(long caffItemId)
+        {
+            var item = await GetCaffItem(caffItemId);
+            return GetFileStream(item.PreviewPath);
+        }
+
+        // Returns a filestream result with Download Name which tells the browser: download the file 
+        public async Task<FileStreamResult> DownloadCaffFile(long caffItemId, long userId)
+        {
+            var item = await GetCaffItem(caffItemId);
+
+            var fileStream = GetFileStream(item.CaffPath);
+            fileStream.FileDownloadName = item.OriginalName;
+            return fileStream;
+        }
+
+        private static FileStreamResult GetFileStream(string path)
+        {
+            new FileExtensionContentTypeProvider().TryGetContentType(path, out var contentType);
+            var stream = File.OpenRead(path);
+            return new FileStreamResult(stream, contentType ?? "application/octet-stream");
         }
     }
 }
